data_source_setup/requirements-source.txt
psycopg2-binary
kafka-python
Faker
faker-commerce
